refactor: kill Timeline.CtxBlocker & register_to_timeline
