+++ janim/anims/display.py
@@ -8,12 +8,12 @@ if TYPE_CHECKING:   # pragma: no cover
 
 class Display(Animation):
     '''
-    在指定的时间区间上显示物件
+    在指定的时间区间上显示物件（不包括后代物件）
     '''
-    def __init__(self, item: 'Item', root_only=False, **kwargs):
+    def __init__(self, item: 'Item', **kwargs):
         super().__init__(**kwargs)
         self.item = item
-        self.root_only = root_only
+        self.timeline.record(item)
 
     def anim_on(self, local_t: float) -> None:
         super().anim_on(local_t)
