+++ janim/anims/transform.py
@@ -1,20 +1,16 @@
 from __future__ import annotations
 
 from collections import defaultdict
-from typing import TYPE_CHECKING, Self
+from typing import Self
 
 from janim.anims.animation import Animation, RenderCall
-from janim.anims.timeline import Timeline
-from janim.components.component import Component
+from janim.components.component import Component, DynamicData
 from janim.constants import ANIM_END_DELTA, OUT
 from janim.items.item import Item
 from janim.logger import log
 from janim.typing import SupportsInterpolate, Vect
 from janim.utils.data import AlignedData
 from janim.utils.paths import PathFunc, path_along_arc, straight_path
-
-if TYPE_CHECKING:
-    from janim.anims.timeline import DynamicData  # pragma: no cover
 
 
 class Transform(Animation):
@@ -81,9 +77,8 @@ class Transform(Animation):
             if tpl in self.aligned:
                 return
 
-            with Timeline.CtxBlocker():
-                data1 = item1.current(as_time=self.global_range.at, skip_dynamic=True)
-                data2 = item2.current(as_time=self.global_range.end, skip_dynamic=True)
+            data1 = item1.current(as_time=self.global_range.at, skip_dynamic=True)
+            data2 = item2.current(as_time=self.global_range.end, skip_dynamic=True)
             aligned = self.aligned[tpl] = data1.align_for_interpolate(data1, data2)
             begin_times[item1] += 1
             end_times[item2] += 1
@@ -148,6 +143,8 @@ class MethodTransform(Transform):
         super().__init__(item, item, **kwargs)
         self.current_alpha = None
 
+        for subitem in item.walk_self_and_descendants():
+            self.timeline.record(subitem)
         self.timeline.detect_changes(item.walk_self_and_descendants())
 
     def do(self, func) -> Self:
@@ -162,8 +159,7 @@ class MethodTransform(Transform):
             aligned = self.aligned[(item, item)]
             alpha = self.get_alpha_on_global_t(global_t)
 
-            with Timeline.CtxBlocker():
-                cmpt = aligned.union.components[key].copy()
+            cmpt = aligned.union.components[key].copy()
             assert isinstance(cmpt, SupportsInterpolate)
 
             cmpt.interpolate(aligned.data1.components[key],
